feat: refactor as independent props
