--- src/containers/BundleContainer.tsx
@@ -1,1 +1,0 @@
-export {};
