--- src/containers/InputContainer.tsx
@@ -1,1 +1,0 @@
-export {};
